function-app/requirements.txt
azure-functions
azure-identity
azure-mgmt-cognitiveservices
azure-mgmt-reservations


